+++ discovery-provider/tests/test_generate_trending.py
